+++ source/renderer/app/actions/app-actions.js
@@ -6,6 +6,8 @@ import Action from './lib/Action';
 export default class AppActions {
   openAboutDialog: Action<any> = new Action();
   closeAboutDialog: Action<any> = new Action();
+  openNetworkStatusDialog: Action<any> = new Action();
+  closeNetworkStatusDialog: Action<any> = new Action();
   getGpuStatus: Action<any> = new Action();
   initAppEnvironment: Action<any> = new Action();
 }