+++ source/renderer/app/containers/Root.js
@@ -13,7 +13,7 @@ export default class Root extends Component<Props> {
   render() {
     const { stores, actions, children } = this.props;
     const { networkStatus, profile, adaRedemption, app, wallets } = stores;
-    const { isNetworkStatusPage, isBlockConsolidationStatusPage } = app;
+    const { isBlockConsolidationStatusPage } = app;
     const { isSettingsPage } = profile;
     const { isAdaRedemptionPage } = adaRedemption;
     const { hasLoadedWallets } = wallets;
@@ -34,7 +34,6 @@ export default class Root extends Component<Props> {
 
     // Just render any page that doesn't require wallets to be loaded or node to be connected
     if (
-      isNetworkStatusPage ||
       (isPageThatDoesntNeedWallets && !isNodeInStoppingSequence) ||
       (isSettingsPage && (isNotEnoughDiskSpace || !isNodeInStoppingSequence))
     ) {